Add files via upload
Introduction :-
The project titled “IMAGE_STEGANOGRAPHY” In large organizations, secrecy is of high importance and maintaining it is a top priority. The intelligence of hackers has made this a very tedious task today. Cryptography was previously used for sending and receiving secret information but it did not prove that secure as it gave suspicions to hackers.
To overcome this problem, steganography came as a solution. It neither gives suspicion to hackers nor affects unintended users. Here, secret information are hidden within data i.e, multimedia data, which is an image, and the data (secret information) can be sent anywhere. The multimedia data simply acts as a medium to maintain secrecy of the information to be conveyed.
In order to hide data within an image, the sender-side program is to be run. And, for receiving the hidden secret information, you need to run the receiver-side program .This digital steganography contains both sender and receiver-side programs. So, user can use a program depending on whether they want to receive or send data.

OBJECTIVE :-
The main objective of  steganography is secret information are hidden within data i.e, multimedia data, which is an image, and the data (secret information) can be sent anywhere.
•	It encrypts all information that you want to send to another person.
•	It saves the time and resources .